Adds information for multiple relying party certificates
